Fix: Move CobraCLI cmd into main
Related Issue #1
